+++ noodles-bcf/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+  * bcf/record/samples: Add series selector (`Samples::select`).
+
 ## 0.50.0 - 2024-04-11
 
 ### Changed